requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
transformers==4.39.3
requests==2.31.0
scikit-image==0.22.0